Renamed operator to sequence, added documentation
